refactoring, traps

+++ src/common.h
@@ -36,8 +36,8 @@
 #include "log.h"
 
 enum        Axis          : int;
-enum        Corner        : int;
-enum class  Side          : int;
+enum        Corner        : int8_t;
+enum class  Side          : int8_t;
 enum class  SpawnClass    : char;
 enum class  InventorySlot : size_t;
 enum class  Element       : size_t;
